Добавляет модуль pristine
