+++ js/api.js
@@ -3,7 +3,7 @@ import { showLoadingDataError, showUploadingDataError } from './errors';
 const BASE_URL = 'https://31.javascript.htmlacademy.pro/kekstagram';
 const Route = {
   GET_DATA: '/data',
-  SEND_DATA: '/2',
+  SEND_DATA: '/',
 };
 const Method = {
   GET: 'GET',
